# cgosqlite: cache column name strings forever