@bradfitz 2023-03-05
This was wrong. It was supposed to be on ColumnName, not ColumnText. But ColumnName isn't hot (or even used) by us currently so not really worth fixing yet.